add dataexport and more math funcs moved
